requirements.txt
Django==1.5
South==0.7.6
dj-database-url==0.2.1
django-mobile==0.2.4
django-twilio==0.4
gunicorn==0.17.2
httplib2==0.7.7
psycopg2==2.4.6
python-memcached==1.48
six==1.2.0
twilio==3.4.3
unittest2==0.5.1
wsgiref==0.1.2
py-pretty==1
